--- apps/server/src/features/users/users.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { UsersService } from './users.service';
-import { UsersController } from './users.controller';
-
-@Module({
-	providers: [UsersService],
-	exports: [UsersService],
-	controllers: [UsersController],
-})
-export class UsersModule {}
