--- apps/server/src/modules/authentication/auth.controller.ts
@@ -1,7 +1,0 @@
-import { Controller, Post, UseGuards, Request, Body } from '@nestjs/common';
-import { AuthService } from './auth.service';
-
-@Controller('auth')
-export class AuthController {
-	constructor(private authService: AuthService) {}
-}
