+++ apps/server/src/modules/authentication/auth.module.ts
@@ -3,10 +3,10 @@ import { AuthService } from './auth.service';
 import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
 import { jwtConstants } from './constants';
-import { UsersModule } from '../../features/users/users.module';
-import { JwtStrategy } from './strategies/jwt.strategy';
-import { AuthController } from './auth.controller';
-import { LocalStrategy } from './strategies/local.strategy';
+import { UsersModule } from '../user/users.module';
+import { JwtStrategy } from './strategy/jwt.strategy';
+import { AuthController } from './controller/auth.controller';
+import { LocalStrategy } from './strategy/local.strategy';
 
 @Module({
 	imports: [
