+++ apps/server/src/modules/authentication/controller/auth.controller.spec.ts
@@ -1,0 +1,26 @@
+import { JwtService } from '@nestjs/jwt';
+import { PassportModule } from '@nestjs/passport';
+import { Test, TestingModule } from '@nestjs/testing';
+import { UsersModule } from '../../user/users.module';
+import { AuthController } from './auth.controller';
+import { AuthService } from '../auth.service';
+import { jwtOptionsProvider } from '../jwt-options.provider';
+import { LocalStrategy } from '../strategy/local.strategy';
+
+describe('AuthController', () => {
+	let controller: AuthController;
+
+	beforeEach(async () => {
+		const module: TestingModule = await Test.createTestingModule({
+			imports: [UsersModule, PassportModule],
+			controllers: [AuthController],
+			providers: [AuthService, LocalStrategy, JwtService, jwtOptionsProvider],
+		}).compile();
+
+		controller = module.get<AuthController>(AuthController);
+	});
+
+	it('should be defined', () => {
+		expect(controller).toBeDefined();
+	});
+});
