+++ apps/server/src/modules/authentication/controller/auth.controller.ts
@@ -1,0 +1,7 @@
+import { Controller, Post, UseGuards, Request, Body } from '@nestjs/common';
+import { AuthService } from '../auth.service';
+
+@Controller('auth')
+export class AuthController {
+	constructor(private authService: AuthService) {}
+}
