+++ apps/server/src/modules/authentication/entity/auth.entity.ts
