+++ apps/server/src/modules/authentication/guard/jwt-auth.guard.ts
