--- apps/server/src/modules/authentication/strategies/jwt.strategy.ts
@@ -1,23 +1,0 @@
-import { ExtractJwt, Strategy } from 'passport-jwt';
-import { PassportStrategy } from '@nestjs/passport';
-import { Injectable } from '@nestjs/common';
-import { jwtConstants } from '../constants';
-import { JwtPayload } from '../interfaces/jwt-payload';
-
-@Injectable()
-export class JwtStrategy extends PassportStrategy(Strategy) {
-	constructor() {
-		super({
-			jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-			ignoreExpiration: false,
-			secretOrKey: jwtConstants.secret,
-			...jwtConstants.jwtOptions,
-		});
-	}
-
-	async validate(payload: JwtPayload) {
-		// TODO check jwt is whitelisted
-		// TODO check user exist/is active
-		return payload;
-	}
-}
