+++ apps/server/src/modules/news/controller/news.controller.spec.ts
@@ -1,0 +1,27 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { NewsController } from './news.controller';
+import { NewsService } from '../news.service';
+import { NewsRepo } from '../repo/news.repo';
+
+describe('NewsController', () => {
+	let controller: NewsController;
+
+	beforeEach(async () => {
+		const module: TestingModule = await Test.createTestingModule({
+			controllers: [NewsController],
+			providers: [
+				NewsService,
+				{
+					provide: NewsRepo,
+					useValue: {},
+				},
+			],
+		}).compile();
+
+		controller = module.get<NewsController>(NewsController);
+	});
+
+	it('should be defined', () => {
+		expect(controller).toBeDefined();
+	});
+});
