+++ apps/server/src/modules/user/entity/user.entity.ts
@@ -1,0 +1,17 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { JwtPayload } from '../../authentication/interface/jwt-payload';
+
+export class UserEntity {
+	constructor(payload: JwtPayload) {
+		// TODO Builder
+		this.userId = payload.userId;
+		this.schoolId = payload.schoolId;
+		this.roles = payload.roles;
+	}
+	@ApiProperty()
+	userId: string;
+	@ApiProperty()
+	schoolId: string;
+	@ApiProperty()
+	roles: string[];
+}
