+++ apps/server/src/shared/core/pipe/parse-object-id.pipe.ts
