+++ apps/server/src/shared/core/repo/entity/base.entity.ts
