+++ apps/server/src/shared/domain/README.md
@@ -1,0 +1,2 @@
+Entities should be moved to a module. They exist only for test reasons as not all legacy entities have been moved to this application. 
+Here: only keep shared / base / abstract classes.
