+++ apps/server/src/shared/domain/entities/school/school.model.ts
@@ -1,0 +1,11 @@
+import { Expose } from 'class-transformer';
+import { ExposeMongoIdAsString } from '../../../core/repo/entity/base.entity';
+import { Types } from 'mongoose';
+
+export class School {
+	@ExposeMongoIdAsString()
+	_id: Types.ObjectId;
+	/** the schools name */
+	@Expose()
+	name: string;
+}
